refactor: modularize rendering with registry
