+++ js/events.js
@@ -2,7 +2,9 @@ import {data} from './data.js';
 import {startStopFight} from './combat.js';
 import {save, exportSave, importSave, hardReset} from './persistence.js';
 import {applyOfflineProgress} from './offline.js';
-import {renderAll, renderStats, renderSkills} from './render.js';
+import {renderAll} from './renderer.js';
+import {renderStats} from './render/stats.js';
+import {renderSkills} from './render/skills.js';
 import {runTests} from './tests.js';
 import {el, clamp} from './utils.js';
 import {showToast} from './toast.js';
